% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_dockerfile}
\alias{create_dockerfile}
\title{Create a Dockerfile with custom options}
\usage{
create_dockerfile(
  r_version = "latest",
  r_packages = c(),
  github_packages = c(),
  base_image = "ubuntu:22.04",
  output_file = "Dockerfile"
)
}
\arguments{
\item{r_version}{The R version to install. Defaults to "latest".}

\item{r_packages}{A character vector of additional CRAN packages to install.}

\item{github_packages}{A character vector of additional GitHub repositories to install.}

\item{base_image}{The base Docker image. Defaults to "ubuntu:22.04".}

\item{output_file}{The name of the Dockerfile to generate. Defaults to "Dockerfile".}
}
\description{
This function generates a Dockerfile for setting up an R environment with specified R and GitHub packages.
}
\examples{
create_dockerfile(r_version = "4.2.3", r_packages = c("ggplot2", "dplyr"), github_packages = c("hadley/devtools"))
}
