% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_build.R
\name{docker_build}
\alias{docker_build}
\title{Build a Docker image}
\usage{
docker_build(
  image_name = "default_image",
  version = "latest",
  build_context = "."
)
}
\arguments{
\item{image_name}{The name of the Docker image. Defaults to "default_image".}

\item{version}{The version of the Docker image. Defaults to "latest".}

\item{build_context}{The build context for the Docker image. Defaults to the current directory.}
}
\description{
This function builds a Docker image with the specified name and version.
}
\examples{
docker_build(image_name = "my_image", version = "1.0")
}
