% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{docker_image_exists}
\alias{docker_image_exists}
\title{Check if a Docker image exists}
\usage{
docker_image_exists(image_name, version = "latest")
}
\arguments{
\item{image_name}{The name of the Docker image.}

\item{version}{The version of the Docker image. Defaults to "latest".}
}
\value{
TRUE if the image exists, FALSE otherwise.
}
\description{
This function checks whether a Docker image with the specified name and version exists.
}
\examples{
docker_image_exists("my_image", "1.0")
}
