% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{docker_list_containers}
\alias{docker_list_containers}
\title{List Docker containers}
\usage{
docker_list_containers(all = FALSE)
}
\arguments{
\item{all}{Logical. If TRUE, includes stopped containers. Defaults to FALSE.}
}
\value{
A character vector of Docker container names and their associated images.
}
\description{
This function lists all available Docker containers. Optionally, it can include stopped containers.
}
\examples{
docker_list_containers()
docker_list_containers(all = TRUE)
}
