% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{docker_list_images}
\alias{docker_list_images}
\title{List Docker images}
\usage{
docker_list_images()
}
\value{
A character vector of Docker images in the format "repository:tag".
}
\description{
This function lists all available Docker images.
}
\examples{
docker_list_images()
}
