% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{docker_prune}
\alias{docker_prune}
\title{Prune unused Docker resources}
\usage{
docker_prune()
}
\description{
This function stops all running containers, removes all containers, and prunes unused images, networks, and volumes.
}
\examples{
docker_prune()
}
