% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_pull.R
\name{docker_pull}
\alias{docker_pull}
\title{Pull a Docker image from Docker Hub}
\usage{
docker_pull(image_name, version = "latest")
}
\arguments{
\item{image_name}{The name of the Docker image.}

\item{version}{The version of the Docker image. Defaults to "latest".}
}
\description{
This function pulls a Docker image from Docker Hub
}
\examples{
docker_pull("my_image", "1.0")
}
