% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{docker_remove}
\alias{docker_remove}
\title{Remove a Docker image}
\usage{
docker_remove(image_name, version = "latest")
}
\arguments{
\item{image_name}{The name of the Docker image.}

\item{version}{The version of the Docker image. Defaults to "latest".}
}
\description{
This function removes a Docker image with the specified name and version if it exists.
}
\examples{
docker_remove("my_image", "1.0")
}
