% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run_condor.R
\name{docker_run_condor}
\alias{docker_run_condor}
\title{Run MFCL Job via HTCondor and Docker}
\usage{
docker_run_condor(
  remote_user,
  remote_host,
  remote_dir,
  github_pat,
  github_username,
  github_org,
  github_repo,
  docker_image,
  target_folder = NULL,
  condor_cpus = NULL,
  condor_memory = NULL
)
}
\arguments{
\item{remote_user}{Character. Remote server username.}

\item{remote_host}{Character. Remote server address.}

\item{remote_dir}{Character. Remote working directory.}

\item{github_pat}{Character. GitHub Personal Access Token.}

\item{github_username}{Character. GitHub username.}

\item{github_org}{Character. GitHub organisation name.}

\item{github_repo}{Character. GitHub repository name.}

\item{docker_image}{Character. Docker image to use.}

\item{target_folder}{Character, optional. Specific folder within the repository to clone (via sparse checkout) 
and where \code{make} is executed. If not provided, the entire repository is used.}

\item{condor_cpus}{Numeric, optional. The number of CPUs to request in the HTCondor job.}

\item{condor_memory}{Character, optional. The amount of memory to request (e.g., "4GB") in the HTCondor job.}
}
\value{
No return value; side effects include transfer of files to the remote server and job submission.
}
\description{
This function clones a GitHub repository (or a specific folder using sparse checkout),
runs the \code{make} command, archives the specified folder (or the entire repository if no
folder is specified) into a tar.gz file, and then creates and submits an HTCondor job on a
remote server using Docker. The function transfers the generated Bash script and Condor
submit file to the remote server before job submission.
}
\details{
The function performs the following steps:

\enumerate{
  \item Creates a fixed Bash script file (\code{run_job.sh}) that:
    \itemizes{
      \item Sets GitHub-related environment variables.
      \item Clones the entire repository. If \code{target_folder} is specified, performs a sparse checkout.
      \item Changes into the appropriate directory and runs \code{make}.
      \item Archives the chosen folder into \code{output_archive.tar.gz} (archives \code{target_folder} if specified, else archives the entire repository).
      \item Cleans up sensitive information.
    }
  \item Creates a HTCondor submit file (\code{condor_job.submit}) that specifies the use of Docker, the executable,
        file transfers, and optionally resource requests such as CPUs and memory.
  \item Checks for the remote directory's existence (creating it if necessary) and then transfers the generated
        script and submit file to the remote server.
  \item Submits the Condor job on the remote server.
}
}
\examples{
\dontrun{
  docker_run_condor(
    remote_user = "myuser",
    remote_host = "remote.server.com",
    remote_dir  = "/home/myuser/jobs",
    github_pat = "ghp_xxxxxxxxxxxxxxxxxxxx",
    github_username = "mygithub",
    github_org = "myorg",
    github_repo = "myrepo",
    docker_image = "mydocker/image:latest",
    target_folder = "src",
    condor_cpus = 4,
    condor_memory = "4GB"
  )
}

}
