% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_stop_container.R
\name{docker_stop_container}
\alias{docker_stop_container}
\title{Stop a Docker container}
\usage{
docker_stop_container(container_id_or_name)
}
\arguments{
\item{container_id_or_name}{The ID or name of the container to stop.}
}
\description{
This function stops a specific Docker container by its ID or name.
}
\examples{
docker_stop_container("my_container")

}
