% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{docker_stop_unnecessary_containers}
\alias{docker_stop_unnecessary_containers}
\title{Stop unnecessary Docker containers}
\usage{
docker_stop_unnecessary_containers(image_name = NULL)
}
\arguments{
\item{image_name}{The name of the Docker image to filter by. Defaults to NULL (all containers).}
}
\description{
This function stops all running Docker containers or filters by a specific image name.
}
\examples{
docker_stop_unnecessary_containers()
docker_stop_unnecessary_containers(image_name = "my_image")
}
