% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run_mfcl2.R
\name{docker_run_mfcl2}
\alias{docker_run_mfcl2}
\title{Run MFCL commands in Docker v2}
\usage{
docker_run_mfcl2(
  image_name,
  commands,
  project_dir = getwd(),
  sub_dirs = NULL,
  parallel = FALSE,
  cores = parallel::detectCores() - 1,
  verbose = TRUE,
  log_file = NULL
)
}
\arguments{
\item{image_name}{The name of the Docker image.}

\item{commands}{A vector of command(s) to execute inside the container.}

\item{project_dir}{The base project directory to mount. Defaults to the current working directory.}

\item{sub_dirs}{A list of subdirectories for execution. Defaults to NULL (only the base directory).}

\item{parallel}{Whether to enable parallel execution. Defaults to FALSE.}

\item{cores}{The number of cores to use for parallel execution. Defaults to all available cores minus one.}

\item{verbose}{Whether to print the executed commands. Defaults to TRUE.}

\item{log_file}{The log file to save the output. Defaults to NULL (no log file).}
}
\value{
A list of results from the executed commands.
}
\description{
This function runs MFCL-related commands inside a Docker container.
}
\examples{
docker_run_mfcl2(image_name = "mfcl_image", commands = "./mfclo64 input_file.frq")

}
